moved code and placed it in the root
