--- amethyst_network/src/net/mod.rs
@@ -1,11 +1,0 @@
-mod net_entity;
-mod net_event;
-mod net_connection;
-mod net_identity;
-mod net_filter;
-
-pub use self::net_entity::NetEntity;
-pub use self::net_event::NetEvent;
-pub use self::net_connection::NetConnection;
-pub use self::net_identity::NetIdentity;
-pub use self::net_filter::NetFilter;
