+++ amethyst_network/src/net_entity.rs
