+++ amethyst_network/src/net_identity.rs
